Weeded out implicit narrowing casts encountered in compiling src/main/[A-Z]*.cpp on 32-bit Linux (as revealed by -Wconversion -Wsign-conversion).
Added CXXR copyright to src/main/edit.cpp .


git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@1466 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/include/CXXR/ConsCell.h
@@ -427,7 +427,7 @@ namespace CXXR {
      */
     inline size_t listLength(const ConsCell* start)
     {
-	return (start ? std::distance(start->begin(), start->end()) : 0);
+	return (start ? size_t(std::distance(start->begin(), start->end())) : 0);
     }
 
     /** @brief <tt>cc ? cc->tail() : 0</tt>
